requirements.txt
behave==1.2.6
playwright==1.40.0
pyyaml==6.0
